Nuevas tablas

+++ src/main/java/com/puntoespresso/puntoespresso/repository/SedeRepository.java
@@ -3,5 +3,5 @@ package com.puntoespresso.puntoespresso.repository;
 import com.puntoespresso.puntoespresso.dto.Sede;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SedeRepository extends JpaRepository<Sede, Long> {
+public interface SedeRepository extends JpaRepository<Sede, Integer> {
 }
